Fixed linter issue, renamed unittest folder to avoid name conflict
Since there is no real reason for this folder to be called "unittest",
I renamed it to "unittests". As a bonus, it's also consistent with the
folder name "tests", both in the plural form.
