+++ tests/oauth2/rfc6749/clients/test_backend_application.py
@@ -5,7 +5,7 @@ from unittest.mock import patch
 from oauthlib import signals
 from oauthlib.oauth2 import BackendApplicationClient
 
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 
 @patch('time.time', new=lambda: 1000)
