+++ tests/oauth2/rfc6749/clients/test_legacy_application.py
@@ -6,7 +6,7 @@ from unittest.mock import patch
 from oauthlib import signals
 from oauthlib.oauth2 import LegacyApplicationClient
 
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 
 @patch('time.time', new=lambda: 1000)
