+++ tests/oauth2/rfc6749/clients/test_service_application.py
@@ -8,7 +8,7 @@ import jwt
 from oauthlib.common import Request
 from oauthlib.oauth2 import ServiceApplicationClient
 
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 
 class ServiceApplicationClientTest(TestCase):
