+++ tests/oauth2/rfc6749/endpoints/test_client_authentication.py
@@ -17,7 +17,7 @@ from oauthlib.oauth2 import (
     RequestValidator, WebApplicationServer,
 )
 
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 from .test_utils import get_fragment_credentials
