+++ tests/oauth2/rfc6749/endpoints/test_extra_credentials.py
@@ -7,7 +7,7 @@ from oauthlib.oauth2 import (
     RequestValidator, WebApplicationServer,
 )
 
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 
 class ExtraCredentialsTest(TestCase):
