+++ tests/oauth2/rfc6749/endpoints/test_introspect_endpoint.py
@@ -5,7 +5,7 @@ from unittest.mock import MagicMock
 from oauthlib.common import urlencode
 from oauthlib.oauth2 import IntrospectEndpoint, RequestValidator
 
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 
 class IntrospectEndpointTest(TestCase):
