+++ tests/oauth2/rfc6749/endpoints/test_resource_owner_association.py
@@ -8,7 +8,7 @@ from oauthlib.oauth2 import (
     RequestValidator, WebApplicationServer,
 )
 
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 from .test_utils import get_fragment_credentials, get_query_credentials
