+++ tests/oauth2/rfc6749/grant_types/test_authorization_code.py
@@ -9,7 +9,7 @@ from oauthlib.oauth2.rfc6749.grant_types import (
 )
 from oauthlib.oauth2.rfc6749.tokens import BearerToken
 
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 
 class AuthorizationCodeGrantTest(TestCase):
