+++ tests/oauth2/rfc6749/grant_types/test_client_credentials.py
@@ -6,7 +6,7 @@ from oauthlib.common import Request
 from oauthlib.oauth2.rfc6749.grant_types import ClientCredentialsGrant
 from oauthlib.oauth2.rfc6749.tokens import BearerToken
 
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 
 class ClientCredentialsGrantTest(TestCase):
