+++ tests/oauth2/rfc6749/grant_types/test_resource_owner_password.py
@@ -9,7 +9,7 @@ from oauthlib.oauth2.rfc6749.grant_types import (
 )
 from oauthlib.oauth2.rfc6749.tokens import BearerToken
 
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 
 class ResourceOwnerPasswordCredentialsGrantTest(TestCase):
