+++ tests/oauth2/rfc8628/clients/test_device.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 from oauthlib import signals
 from oauthlib.oauth2 import DeviceClient
 
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 
 class DeviceClientTest(TestCase):
