+++ tests/openid/connect/core/endpoints/test_claims_handling.py
@@ -12,7 +12,7 @@ from oauthlib.openid import RequestValidator
 from oauthlib.openid.connect.core.endpoints.pre_configured import Server
 
 from tests.oauth2.rfc6749.endpoints.test_utils import get_query_credentials
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 
 class TestClaimsHandling(TestCase):
