+++ tests/openid/connect/core/endpoints/test_userinfo_endpoint.py
@@ -5,7 +5,7 @@ from unittest import mock
 from oauthlib.oauth2.rfc6749 import errors
 from oauthlib.openid import RequestValidator, UserInfoEndpoint
 
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 
 def set_scopes_valid(token, scopes, request):
