+++ tests/openid/connect/core/grant_types/test_base.py
@@ -5,7 +5,7 @@ from unittest import mock
 from oauthlib.common import Request
 from oauthlib.openid.connect.core.grant_types.base import GrantTypeBase
 
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 
 class GrantBase(GrantTypeBase):
