+++ tests/test_uri_validate.py
@@ -2,7 +2,7 @@ import unittest
 
 from oauthlib.uri_validate import is_absolute_uri
 
-from tests.unittest import TestCase
+from tests.unittests import TestCase
 
 
 class UriValidateTest(TestCase):
